+++ frontend/src/App.tsx
@@ -4,9 +4,9 @@ import React, { useState } from 'react';
 import { BrowserRouter as Router, Navigate, Outlet, Route, Routes, useLocation } from 'react-router-dom';
 import './App.css';
 import { Dashboard } from './Components/Dashboard/Dashboard';
+import DatasetMetadataTable from './Components/DatasetMetadataTable/DatasetMetadataTable';
 import { MapContextForm } from './Components/MapContextForm/MapContextForm';
 import MapContextTable from './Components/MapContextTable/MapContextTable';
-import DatasetMetadataTable from './Components/Metadata/DatasetMetadataTable';
 import { NavBar } from './Components/NavBar/NavBar';
 import FeatureTypeTable from './Components/OgcService/FeatureTypeTable';
 import LayerTable from './Components/OgcService/LayerTable';
